backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
paho-mqtt
httpx
alembic
python-dotenv
